gc7.0alpha5 tarball import

+++ headers.c
@@ -254,7 +254,7 @@ struct hblkhdr * GC_install_header(struct hblk *h)
     result = alloc_hdr();
     SET_HDR(h, result);
 #   ifdef USE_MUNMAP
-	result -> hb_last_reclaimed = GC_gc_no;
+	result -> hb_last_reclaimed = (unsigned short)GC_gc_no;
 #   endif
     return(result);
 }
